feat: initial commit
Yes, this code is _terrible._ Fortunately, I don't care! The objective
right now is to make something that works, not something that is good.
This is being created to be used in another project, not as a standalone
thing itself. This is only standalone because I expect to use it in
other projects for myself. I expect I will target a higher standard for
quality later, but right now the objective is to move as fast as I can
to make something that works pretty good (I have too many unfinished
projects to shoot for the stars on _another_).