build(licenses): harvest license texts from `choosealicense.com`
Added a script (and an initial pull) to pull license texts from
[`choosealicense.com`](https://github.com/github/choosealicense.com).
I'm going to print license texts in Markdown code blocks, so the very
long lines of text in the texts embedded by the `spdx` crates would be
obnoxious. The license texts from `choosealicense.com` are hand-wrapped
to 78 columns, so they should be much more readable.